refactor: switches to self-closing tags
